+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - No more PHP containers
 - Storage management v2
 - backup and restore module
+- Delete B603 analysis
 
 ## [3.0.1] - 2021-03-09
 ### Added